test(Overlay): update test cases
